% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_game_outcomes.R
\name{add_game_outcomes.environment}
\alias{add_game_outcomes.environment}
\title{Add game outcomes using the environment method}
\usage{
add_game_outcomes.environment(prisoner_game)
}
\arguments{
\item{prisoner_game}{prisoner_game. A game created by \verb{create prisoner_game}}
}
\description{
Add game outcomes using the environment method
}
