% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{create_outcomes}
\alias{create_outcomes}
\title{Create game outcomes based on a method}
\usage{
create_outcomes(rooms_ls, pick_max, prisoners_n, method)
}
\arguments{
\item{rooms_ls}{list. A list of rooms created by \code{create_rooms}}

\item{pick_max}{integer. The maximum number of boxes that can be opened}

\item{method}{character. The method to use. Can be 'environment' or 'agent'}
}
\description{
Create game outcomes based on a method
}
