% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{create_outcomes_agent}
\alias{create_outcomes_agent}
\title{Add game outcomes using the agent method}
\usage{
create_outcomes_agent(rooms_ls, prisoners_n, pick_max)
}
\arguments{
\item{prisoner_game}{prisoner_game. A game created by \verb{create prisoner_game}}
}
\description{
Add game outcomes using the agent method
}
