% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcomes.R
\name{create_outcomes_environment}
\alias{create_outcomes_environment}
\title{Add game outcomes using the environment method}
\usage{
create_outcomes_environment(rooms, prisoners_n, open_max)
}
\arguments{
\item{rooms}{list. A list of rooms created by \code{create_rooms}}

\item{prisoners_n}{integer. The number of prisoners}

\item{open_max}{integer. The maximum number of boxes that can be opened}
}
\description{
Add game outcomes using the environment method
}
