% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{create_rooms}
\alias{create_rooms}
\title{Title}
\usage{
create_rooms(boxes_n = 100, rooms_n = 1)
}
\arguments{
\item{boxes_n}{integer. The number of boxes in the room}

\item{rooms_n}{integer. The number of rooms}
}
\description{
Title
}
