% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_room}
\alias{plot_room}
\title{Title}
\usage{
plot_room(room, color_by_loop = TRUE, sort_by_loop = TRUE, show_loops = TRUE)
}
\arguments{
\item{room}{room. A room object created by \code{create_room()}}

\item{color_loops}{logical. Should the boxes be colored by loop?}

\item{arrange_loops}{logical. Should the boxes be arranged?}
}
\description{
Title
}
