% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{prisoners_create_game}
\alias{prisoners_create_game}
\title{Create a new prisoner game object}
\usage{
prisoners_create_game(teams_n, prisoners_n, open_max, method = "environment")
}
\arguments{
\item{teams_n}{integer. The number of teams of prisoners}

\item{prisoners_n}{integer. The number of prisoners per team}

\item{open_max}{integer. The maximum number of boxes individual prisoners can open}

\item{method}{character. The method to use for simulation. Either "agent" or "environment"}
}
\value{
list. A list with the structure of the game
}
\description{
Create a new prisoner game object
}
