% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_room.R
\name{prisoners_create_room}
\alias{prisoners_create_room}
\title{Title}
\usage{
prisoners_create_room(boxes_n = 100)
}
\arguments{
\item{boxes_n}{integer. The number of boxes in the room}
}
\description{
Title
}
