% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisoners_simulate.R
\name{prisoners_simulate}
\alias{prisoners_simulate}
\title{Simulate the prisoner game}
\usage{
prisoners_simulate(
  prisoners_n = 100,
  teams_n = 100,
  pick_max = floor(prisoners_n/2),
  method = c("environment", "agent")
)
}
\arguments{
\item{prisoners_n}{integer. The number of prisoners in each team}

\item{teams_n}{integer. The number of (separate) teams of prisoners}

\item{pick_max}{integer. The maximum number of boxes individual prisoners can open}

\item{method}{character. The method to use for simulation. Either "agent" or "environment"}
}
\value{
list. A list with the structure of the game
}
\description{
Simulate the prisoner game
}
