% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull_duration_total}
\alias{pull_duration_total}
\title{Pull the success rate of the prisoners game}
\usage{
pull_duration_total(game)
}
\arguments{
\item{game}{A game object}
}
\description{
Pull the success rate of the prisoners game
}
