% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull_room}
\alias{pull_room}
\title{Pull a room out of a list of rooms}
\usage{
pull_room(room_ls, room)
}
\arguments{
\item{room_ls}{list. A list of rooms created by \code{create_rooms()}}

\item{room}{integer. The room number to pull}
}
\value{
tbl. A tbl indicating a room
}
\description{
Pull a room out of a list of rooms
}
