% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull.R
\name{pull_success_p}
\alias{pull_success_p}
\title{Pull the success rate of the prisoners game}
\usage{
pull_success_p(game)
}
\arguments{
\item{game}{A game object}
}
\description{
Pull the success rate of the prisoners game
}
