% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_agent.R
\name{simulate_agent_prisoner}
\alias{simulate_agent_prisoner}
\title{Title}
\usage{
simulate_agent_prisoner(room = NULL, prisoner_i = 1, pick_max = NULL)
}
\arguments{
\item{room}{tbl. A tibble created by prisoners_create_room()}

\item{prisoner_i}{integer. The prisoner number}

\item{pick_max}{integer. The maximum number of boxes to search}
}
\value{
list
}
\description{
Title
}
