% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_agent.R
\name{simulate_agent_team}
\alias{simulate_agent_team}
\title{Simulate the Prisoners and Boxes Game}
\usage{
simulate_agent_team(room, pick_max = NULL)
}
\arguments{
\item{prisoners_n}{integer. The number of prisoners}
}
\value{
list. A list with the following elements:
\itemize{
\item is_success: logical. Whether all prisoners were successful
\item boxes_opened_n_vec: integer vector. The number of boxes opened by each prisoner
\item team_result_ls: list. The results of each prisoner's search
}
}
\description{
Simulate the Prisoners and Boxes Game
}
